Added OpenGL support.
Also refactored the DX12 implementation, added more documentation.

We could use some additions that make it easier to unit test and communicate shared dependencies. We don't include as a submodule CrossWindow and CrossWindowGraphics, but some way of cloning those and including them would be nice for testing. `XGFX_IMGUI_TEST=On`.

Fixed formatting.